chore (refactor): refactor code
- add controllers
- clean up code

[Maintains #]
